requirements.txt
sqlalchemy==2.0.23
passlib==1.7.4
psycopg[binary]==3.1.15
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1 # passlib doesn't support 4.1

# Testing
pytest==7.4.2
pytest-mock==3.12.0

